Added share import and theme changes

+++ app/src/main/java/com/osito/acordes/EditActivity.kt
@@ -1,5 +1,7 @@
 package com.osito.acordes
 
+import android.app.AlertDialog
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
@@ -21,16 +23,39 @@ class EditActivity : AppCompatActivity() {
             binding.content.setText(readFromFile(file!!))
         }
 
+        if (intent.hasExtra(Intent.EXTRA_STREAM)) {
+            val inputStream = contentResolver.openInputStream(intent.getParcelableExtra(Intent.EXTRA_STREAM)!!)
+            val text = inputStream!!.bufferedReader().use { it.readText() }
+            binding.content.setText(text)
+        }
+
         binding.save.setOnClickListener {
             Toast.makeText(this, "Archivo Guardado", Toast.LENGTH_SHORT).show()
-            writeToFile(binding.title.text.toString(), binding.content.text.toString().replace("&nbsp!", " "))
+            var text = binding.content.text.toString()
+            text = text
+                .replace("&nbsp!", " ")
+                .replace("\r", "")
+            writeToFile(
+                binding.title.text.toString(),
+                text
+            )
             finish()
         }
 
         binding.delete.setOnClickListener {
-            Toast.makeText(this, "Archivo Eliminado", Toast.LENGTH_SHORT).show()
-            File(filesDir.absolutePath + "/" + binding.title.text).delete()
-            finish()
+            AlertDialog.Builder(this).apply {
+                setMessage("Seguro que quiere eliminar el archivo?")
+                setPositiveButton("Si") { _, _ ->
+                    Toast.makeText(this@EditActivity, "Archivo Eliminado", Toast.LENGTH_SHORT).show()
+                    File(filesDir.absolutePath + "/" + binding.title.text).delete()
+                    finish()
+                }
+                setNegativeButton("No") { dialog, _ ->
+                    dialog.dismiss()
+                }
+            }
+                .create()
+                .show()
         }
     }
